+++ packages/tailwindcss/src/utilities.test.ts
@@ -11825,6 +11825,7 @@ test('bg-clip', async () => {
       }
 
       .bg-clip-text {
+        -webkit-background-clip: text;
         background-clip: text;
       }"
     `)